+++ config/routes.rb
@@ -1,5 +1,14 @@
 Rails.application.routes.draw do
-  root 'items#index'
+  root 'items#spreadsheet'
 
-  resources :items
+  resources :items do
+    collection do
+      get :spreadsheet
+      post :add_row
+    end
+
+    member do
+      patch :update_cell
+    end
+  end
 end