+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 1.0.0
 
 ### Enhancements
-- Items can now be created
+- Items can now be created on a spreadsheet
   - `box`: The current box they are in
   - `name`: The name of the item
   - `location`: The place where the item should end up (kitchen, bedroom, etc)